requirements.txt
flask==2.3.3
Flask-SQLAlchemy~=3.1
Flask-Migrate==4.0.5
Jinja2==3.1.3

#dev dependencies
python-dotenv~=1.0.0
alembic~=1.13.1
SQLAlchemy~=2.0.23
